+++ Evence/Components/index.js
@@ -1,0 +1,6 @@
+
+
+export {default as Test} from './Test'
+// export {default as Home} from './Home'
+
+// console.log('Home component in the index', Home)
